third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-fonts/fallback-url-to-local-expected.txt

This is a testharness.js-based test.
[FAIL] We should use the local font to render the page when the primary remote font is loading
  assert_equals: expected 250 but got 150
Harness: the test ran to completion.
